backend/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
defusedxml==0.7.1
Django==4.1.5
django-rest-framework==0.1.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
docopt==0.6.2
idna==3.4
install==1.3.5
pipreqs==0.4.11
psycopg2-binary==2.9.5
pycparser==2.21
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.3
urllib3==1.26.14
yarg==0.1.9
python-dotenv==0.21.0
Pillow==9.4.0
django-filter==22.1
